Frontend adjustments #Tali

+++ src/main/java/com/example/NutriTrack/repo/FoodConsumptionRepository.java
@@ -4,6 +4,8 @@ package com.example.NutriTrack.repo;
 import com.example.NutriTrack.model.FoodConsumption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodConsumptionRepository extends JpaRepository<FoodConsumption, Long> {
+import java.util.UUID;
+
+public interface FoodConsumptionRepository extends JpaRepository<FoodConsumption, Integer> {
     // Add custom query methods if needed
 }
